requirements.txt
pytest
selenium
allure-pytest
python-dotenv